cms/requirements.txt
Flask>=3.1,<4
Markdown>=3.7,<4
PyYAML>=6.0,<7
bleach>=6.2,<7
